en saa nyt oikein @secured toimimaan niin tässä on kömpelö ratkaisu

+++ src/main/java/wepa/controller/TunnisteController.java
@@ -18,14 +18,16 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import wepa.domain.Tunniste;
+import wepa.service.LoggedInKayttajaService;
 import wepa.service.TunnisteService;
 
 @Controller
 public class TunnisteController {
 
     @Autowired
+    LoggedInKayttajaService loggedInKayttajaService;
+    @Autowired
     TunnisteService tunnisteService;
-
 
     //Haetaan tunnisteet
     @RequestMapping(value = "/tunnisteet", method = RequestMethod.GET)
@@ -35,12 +37,11 @@ public class TunnisteController {
         return "tunnisteet";
     }
 
-
     //Luodaan uusi tunniste
     @RequestMapping(value = "/tunnisteet", method = RequestMethod.POST)
     public String lisaaTunniste(@Valid @ModelAttribute Tunniste tunniste,
             BindingResult bindingResult, Model model) {
-        if(bindingResult.hasErrors()){
+        if (bindingResult.hasErrors()) {
             model.addAttribute("tunniste", tunniste);
             model.addAttribute("tunnisteet", tunnisteService.findAll());
             return "tunnisteet";
@@ -50,37 +51,36 @@ public class TunnisteController {
         }
         return "redirect:/tunnisteet";
     }
-    
+
     //Liitetään tunniste kuvaan
-    @RequestMapping(value="/{kuvaId}/lisaatunniste", method = RequestMethod.POST)
+    @RequestMapping(value = "/{kuvaId}/lisaatunniste", method = RequestMethod.POST)
     public String lisaaTunnisteKuvaan(@PathVariable Long kuvaId, @RequestParam Long tunnisteId) {
         tunnisteService.lisaatunnisteKuvaan(kuvaId, tunnisteId);
-        return "redirect:/pics/"+kuvaId;
+        return "redirect:/pics/" + kuvaId;
     }
-    
 
     //Poistetaan tunniste kuvasta
     @Secured("ADMIN")
-    @RequestMapping(value="/{kuvaId}/poistatunniste/{tunnisteId}", method = RequestMethod.DELETE)
+    @RequestMapping(value = "/{kuvaId}/poistatunniste/{tunnisteId}", method = RequestMethod.DELETE)
     public String poistaTunnisteKuvasta(@PathVariable Long kuvaId, @PathVariable Long tunnisteId) {
         tunnisteService.poistaTunnisteKuvasta(kuvaId, tunnisteId);
-        return "redirect:/pics/"+kuvaId;
+        return "redirect:/pics/" + kuvaId;
     }
-    
+
     // Haetaan tunnisteeseen liittyvät kuvat
     @RequestMapping(value = "/tunnisteet/{tunnisteId}", method = RequestMethod.GET)
     public String getTunniste(Model model, @PathVariable Long tunnisteId) {
         model.addAttribute("tunniste", tunnisteService.findOne(tunnisteId));
         return "tunniste";
     }
-    
-    
 
     //Poistetaan tunniste kokonaan
-    @Secured("ADMIN")
-    @RequestMapping(value="/tunnisteet/{tunnisteId}", method = RequestMethod.DELETE)
+//    @Secured("ADMIN")
+    @RequestMapping(value = "/tunnisteet/{tunnisteId}", method = RequestMethod.DELETE)
     public String deleteTunniste(@PathVariable Long tunnisteId) {
-        tunnisteService.delete(tunnisteId);
+        if (this.loggedInKayttajaService.getAuthenticatedKayttaja().getAuthority().equals("ADMIN")) {
+            tunnisteService.delete(tunnisteId);
+        }
         return "redirect:/tunnisteet";
     }
